Organize scalacheck.Arbitrary instances in quasar-core
Where appropriate, extracts/renames Arbitrary instances
into their own files for easy a la carte importing and adds
aggregate 'Arbitraries' objects to enable easy importing of
all the instances in a package.

Also adds a few arbitrary instances for types in
quasar.fs.mount
