+++ core/src/test/scala/quasar/Arbitraries.scala
@@ -1,0 +1,7 @@
+package quasar
+
+trait Arbitraries extends
+  DataArbitrary with
+  TypeArbitrary
+
+object Arbitraries extends Arbitraries
