+++ core/src/test/scala/quasar/CoreArbitraries.scala
@@ -1,0 +1,10 @@
+package quasar
+
+trait CoreArbitraries extends
+  Arbitraries with
+  config.Arbitraries with
+  fs.Arbitraries with
+  fs.mount.Arbitraries with
+  sql.Arbitraries
+
+object CoreArbitraries extends CoreArbitraries
