+++ core/src/test/scala/quasar/config/Arbitraries.scala
@@ -1,0 +1,6 @@
+package quasar.config
+
+trait Arbitraries extends
+  CoreConfigArbitrary
+
+object Arbitraries extends Arbitraries
