+++ core/src/test/scala/quasar/fs/mount/Arbitraries.scala
@@ -1,0 +1,8 @@
+package quasar.fs.mount
+
+trait Arbitraries extends
+  MountConfigArbitrary with
+  ConnectionUriArbitrary with
+  MountsArbitrary
+
+object Arbitraries extends Arbitraries
