+++ core/src/test/scala/quasar/fs/mount/ConnectionUriArbitrary.scala
@@ -1,0 +1,15 @@
+package quasar.fs.mount
+
+import quasar.Predef._
+
+import org.scalacheck.{Arbitrary, Gen}
+
+trait ConnectionUriArbitrary {
+  implicit val connectionUriArbitrary: Arbitrary[ConnectionUri] =
+    Arbitrary(for {
+      scheme <- Gen.alphaStr
+      rest   <- Gen.alphaStr
+    } yield ConnectionUri(s"$scheme:$rest"))
+}
+
+object ConnectionUriArbitrary extends ConnectionUriArbitrary
