+++ core/src/test/scala/quasar/sql/Arbitraries.scala
@@ -1,0 +1,6 @@
+package quasar.sql
+
+trait Arbitraries extends
+  ExprArbitrary
+
+object Arbitraries extends Arbitraries
