+++ it/src/test/scala/quasar/fs/ReadFilesSpec.scala
@@ -1,7 +1,7 @@
 package quasar.fs
 
 import quasar.Predef._
-import quasar.{Data, DataGen}
+import quasar.Data
 import quasar.fp._
 
 import java.lang.RuntimeException
